+++ astropy/stats/histogram.py
@@ -8,7 +8,6 @@ Ported from the astroML project: http://astroML.org/
 
 import numpy as np
 from . import bayesian_blocks
-from . import median_absolute_deviation
 
 __all__ = ['histogram', 'scott_bin_width', 'freedman_bin_width',
            'knuth_bin_width']
@@ -207,12 +206,18 @@ def freedman_bin_width(data, return_bins=False):
 
     if return_bins:
         dmin, dmax = data.min(), data.max()
-        if dx < 1e-6:
-            dx = median_absolute_deviation(data)
-        if dx < 1e-6:
-            raise ValueError("data has too small IQR and median_absolute_deviation values")
         Nbins = max(1, np.ceil((dmax - dmin) / dx))
-        bins = dmin + dx * np.arange(Nbins + 1)
+        try:
+            bins = dmin + dx * np.arange(Nbins + 1)
+        except ValueError as e:
+            if 'Maximum allowed size exceeded' in str(e):
+                raise ValueError(
+                    'The inter-quartile range of the data is too small: '
+                    'failed to construct histogram with {} bins. '
+                    'Please use another bin method, such as '
+                    'bins="scott"'.format(Nbins + 1))
+            else:  # Something else  # pragma: no cover
+                raise
         return dx, bins
     else:
         return dx