+++ CHANGES.rst
@@ -966,9 +966,8 @@ astropy.samp
 astropy.stats
 ^^^^^^^^^^^^^
 
-- In ``freedman_bin_width``, if the data has too small IQR, use median
-  absolute deviation instead. If both IQR and median absolute deviation
-  values are too small, give an value error. [#7248]
+- In ``freedman_bin_width``, if the data has too small IQR,
+  raise ``ValueError``. [#7248, #7402]
 
 astropy.table
 ^^^^^^^^^^^^^